Move live exchange-rate tests out of unit suite
